requirements.txt
fastapi
uvicorn
sqlalchemy
databases
bcrypt
python-jose[cryptography]  
requests                  
httpx                    
python-dotenv              
pytest
pytest-asyncio
pydantic[email]
passlib
bcrypt
aiosqlite
pydantic_settings
